modify: s3Utils.js
